Split task creation for Android and Jvm only projects
