+++ apollo-gradle-plugin/build.gradle
@@ -15,6 +15,7 @@ dependencies {
 
   implementation localGroovy()
   implementation project(':apollo-compiler')
+  implementation dep.guavaJre
   implementation dep.gradleNodePlugin
   implementation dep.moshi
