+++ apollo-gradle-plugin/src/main/groovy/com/apollographql/apollo/gradle/android/AndroidTaskConfiguratorFactory.groovy
@@ -1,0 +1,10 @@
+package com.apollographql.apollo.gradle.android
+
+import org.gradle.api.Project
+
+class AndroidTaskConfiguratorFactory {
+
+    static AndroidTaskConfigurator create(Project project) {
+        return new AndroidTaskConfigurator(project)
+    }
+}
