+++ apollo-gradle-plugin/src/main/groovy/com/apollographql/apollo/gradle/jvm/JvmTaskConfiguratorFactory.groovy
@@ -1,0 +1,10 @@
+package com.apollographql.apollo.gradle.jvm
+
+import org.gradle.api.Project
+
+class JvmTaskConfiguratorFactory {
+
+    static JvmTaskConfigurator create(Project project) {
+        return new JvmTaskConfigurator(project)
+    }
+}
