+++ gradle/dependencies.gradle
@@ -36,6 +36,7 @@ ext.dep = [
     recyclerView            : "com.android.support:recyclerview-v7:$versions.supportLibVersion",
     compiletesting          : 'com.google.testing.compile:compile-testing:0.15',
     cache                   : "com.nytimes.android:cache:$versions.cacheVersion",
+    guavaJre                : 'com.google.guava:guava:28.0-jre',
     javaPoet                : 'com.squareup:javapoet:1.9.0',
     kotlinpoet              : 'com.squareup:kotlinpoet:1.3.0',
     moshi                   : "com.squareup.moshi:moshi:$versions.moshi",
